下一步：给Window类增加创建窗口的能力，而不只是绑定窗口，还有解决最早遗留下来的问题，如窗口标题，图标，资源无法打包在EXE内等。

+++ DUI_Demo/ControlBase.cpp
@@ -174,8 +174,8 @@ BOOL ControlBase::GetCursorPos(Point * pt)
 {
 	BOOL ret;
 	ret = m_Parent->GetCursorPos(pt);
-	pt->X -= m_Rect->X;
-	pt->Y -= m_Rect->Y;
+	pt->X -= (INT)m_Rect->X;
+	pt->Y -= (INT)m_Rect->Y;
 	return ret;
 }
